+++ pages/organizations/_organizationId/forms/index.vue
@@ -105,7 +105,8 @@ export default {
     },
     async del(id) {
       let r = confirm('Are you sure you want to delete the data?')
-      await this.$axios.delete(`/forms/${id}`)
+      if(r)
+        await this.$axios.delete(`/forms/${id}`)
       this.getData()
     }
   }
